don't leave a dir marked nonexistent after creating it
2002-06-02  Havoc Pennington  <hp@pobox.com>

	* backends/xml-cache.c (cache_lookup): don't leave a dir marked
	nonexistent after creating it
	(cache_set_nonexistent): remove hash entry _before_ freeing
	the current key. Should fix assorted "gconf loses my settings"
	kind of bugs.

	* backends/xml-dir.c (dir_sync): don't try to delete the root dir,
	should fix some warnings.

+++ backends/xml-dir.c
@@ -374,11 +374,14 @@ dir_sync (Dir      *d,
           return FALSE;
         }
 
-      if (rmdir (d->fs_dirname) != 0)
+      if (strcmp (d->key, "/") != 0) /* don't delete root dir */
         {
-          gconf_set_error (err, GCONF_ERROR_FAILED, _("Failed to delete \"%s\": %s"),
-                           d->fs_dirname, strerror (errno));
-          return FALSE;
+          if (rmdir (d->fs_dirname) != 0)
+            {
+              gconf_set_error (err, GCONF_ERROR_FAILED, _("Failed to delete \"%s\": %s"),
+                               d->fs_dirname, strerror (errno));
+              return FALSE;
+            }
         }
 
       if (deleted)
@@ -562,7 +565,7 @@ dir_get_value   (Dir* d,
                  GError** err)
 {
   Entry* e;
-
+  
   if (d->doc == NULL)
     dir_load_doc(d, err);
